{{cookiecutter.project_slug}}/requirements.txt
# =================
# core requirements
# =================
Sphinx==7.3.7
selenium==4.21.0
locust==2.28.0
safety==3.2.0
unittest-parallel==1.6.1
retry==0.9.2
click==8.1.7
jinja2==3.1.4
pytest==8.2.1
pytest-asyncio==0.23.7
python-dotenv==1.0.1
scour==0.38.2
sqlalchemy==2.0.31
requests==2.32.2
# === Database options ===
psycopg2-binary==2.9.9
# === Database diagnostics in core.adapters.database ===
pandas==2.2.2
matplotlib==3.9.0
# === TODO advanced editors - for core.editors ===
imageio==2.34.1
python-pptx==0.6.23
pillow==10.3.0
svglib==1.5.1
reportlab==4.2.0
PyPDF2==3.0.1
pdf2image==1.17.0
# === TODO NLP tools - for core.adapters.nlp ===
joblib==1.4.2
gensim==4.3.2
pyspellchecker==0.8.1
scikit-learn==1.5.0
# # === TODO research tools - for core.adapters.research ===
wikipedia==1.4.0
polygon==1.2.0
fredapi==0.5.2
# # === TODO scraping tools - for core.adapters.scraping ===
beautifulsoup4==4.12.3
newspaper4k==0.9.3.1
# # === TODO social tools - for core.adapters.research.social ===
praw==7.7.1
# === TODO AWS Cloud services - for core.adapters.storage.storage_aws ===
boto3==1.34.113
botocore==1.34.113
moto==5.0.9
# === TODO GCP Cloud Storage services - for core.adapters.storage.storage_gcp ===
google-cloud-storage==2.16.0
google-cloud-secret-manager==2.20.0
# === TODO GCP Cloud Big Query services - for core.adapters.storage.storage_gcp ===
pandas-gbq==0.23.1
# # === TODO Comms services - for core.adapters.alert.alert_push_slack ===
slack_sdk==3.29
# # === TODO PubSub services - for core.adapters.pubsub.redis ===
redis==5.0.6
# # === TODO LLM services - for core.adapters.llms.agent ===
pydantic==2.7.4
langchain==0.2.5
langchain-community==0.2.5
openai==1.34.0
# # === TODO serial controller - for core.adapters.mechatronics ===
pyserial==3.5

# ================
# api requirements
# ================
# # === webserver services ===
Flask==2.2.5
flask-cors==4.0.1
flask-ipban==1.1.5
# Flask-Login==0.6.3
Flask-Migrate==4.0.7
# Flask-Minify==0.42
Flask-SQLAlchemy==3.1.1
flasgger==0.9.5
Flask-Limiter==3.7.0
flask-talisman==1.1.0
Flask-JWT-Extended==4.3.1
gunicorn==22.0.0
marshmallow==3.21.3
passlib==1.7.4
werkzeug==3.0.3
